Reverse String
Firstly I started my line of code with the original string hello. I then inputted the reversed string and used the system.out.println function in this line of this code. I tried and experimented with different lines of code to get different results as well. I'm starting to realize the relation to Java and the methods needed to succeed.