Add files via upload
I start my line of code with the main method as the entry point then public static int that utilizes the comparetwoints method that uses int a and int b as parameters. It also utilizes the math max function to find and return the largest of the two integers. The sumtwoints takes the two integers a and b and calculates and then returns the sum of the integers. The integer variables num 1 and num 2  that are 24 and 23 which are their values. The values of num 1 and num 2 are then printed. The comparetwoints then includes num 1, and num 2 and the result being stored in the largerint variable the larger integer is next printed. The sum twoints is followed with num 1 and num 2 with the result being stored in the sum variable. The sum is printed.