Add files via upload
I first start with a prompt and enter three integers to store the largest of the three integers. Then continue with an if statement that uses num 2 being less than num 1 showing that if it is true that num 1 would be greater than or equal to num 2. The next part is that if num 2 is greater than or equal to num 3 and if it is true that it would assign the value of num 3 to the largest. I then come to an end with a system print Ln that asks for the largest integer.